pyproject.toml
[tool.poetry]
name = "shops"
version = "0.1.0"
description = "Simple API service based on Django REST Framework."
authors = ["Pavel Borisov <pavel.borisofff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.10.1"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
